requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.5
django-mongodb-engine==0.6.0
djangorestframework==3.14.0
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.4.2
et-xmlfile==1.1.0
openpyxl==3.1.2
pymongo==4.5.0
pytz==2023.3.post1
sqlparse==0.2.4
typing-extensions==4.8.0
